travis: run tests/ and api_tests/
/!\ tests must run in this order when run all together

For simplicity, the current implem of cli tools scenarios
in `tests/` expects that no other experiment is underway
and does not like another experiment being around in its back.

Tests in `api_tests/` create several experiments that last a bit
after the tests are finished, causing `test_cli_tools.py` to
fail when run immediately afterwards, for lack of `-i exp_id`
in the command-line arguments.